feature15 updated according to accepted architecture
