+++ src/bot/constants/state.py
@@ -1,2 +1,4 @@
 CHECK: int = 1
-MENU: int = 2
+MAIN_MENU: int = 2
+FIO: int = 9
+
